% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateQTI.R
\name{create_groupxml}
\alias{create_groupxml}
\title{Create the XML text for a group of quiz questions}
\usage{
create_groupxml(df, groupi, respids)
}
\arguments{
\item{df}{data.frame where each row is a question. The column names should be the following:
\itemize{
\item G: Question group
\item 'Question Type': Character string identifying the type of question, should equal either 'MC' for a multiple choice question or 'Essay' for an essay/short-answer question
\item 'Text Type': Character string identifying whether the question text is plain text or html, should equal either 'html' or 'plain'
\item Points: Number of points for the question
\item Question: Text of the question
\item A: Correct answer, a number 1 thru 4
\item 'Choice 1': Choice 1
\item 'Choice 2': Choice 2
\item 'Choice 3': Choice 3
\item 'Choice 4': Choice 4
\item Feedback: General feedback given to students after they complete the quiz and answers are shown
}}

\item{groupi}{Character string of the group to create, which must be exactly one of the strings in column 'G' in the data frame \code{df}}

\item{respids}{Vector of unique four-digit integers with length equal to the number of choices in the whole quiz. For example, if there are 10 questions in the quiz and each question has 4 answer choices, \code{respids} will be a vector of length 10*4 = 40.}
}
\value{
Character string of the XML code for a group of multiple-choice quiz questions
}
\description{
This function creates the XML code for a group of questions, including the XML creating the group and the XML for each individual question.
}
