% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateQTI.R
\name{create_longid}
\alias{create_longid}
\title{Create an identity for a QTI item.}
\usage{
create_longid(size = 33)
}
\arguments{
\item{size}{Number of digits for the hex code. Defaults to 33.}
}
\value{
Character string of lenghth 34, starting with the letter 'g' and followed by the 33-digit hex code, because that's what Canvas wants?
}
\description{
Each item in a QTI file contains a unique 33 digit hex code. This function just creates a random 33 character hex code. There are 5.4 duodecillion unique 33-digit hex codes, so it's extremely unlikely two will be the same.
}
