% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateQTI.R
\name{create_questionxml}
\alias{create_questionxml}
\title{Create the XML text for a single multiple-choice question}
\usage{
create_questionxml(df, qn, respids)
}
\arguments{
\item{df}{data.frame where each row is a question. The column names should be the following:
\itemize{
\item G: Question group
\item Question: Text of the question
\item A: Correct answer, a number 1 thru 4
\item 'Choice 1': Choice 1
\item 'Choice 2': Choice 2
\item 'Choice 3': Choice 3
\item 'Choice 4': Choice 4
\item Feedback: General feedback given to students after they complete the quiz and answers are shown
}}

\item{qn}{integer identifying the row in \code{df} that the individual question of interest appears}

\item{respids}{Vector of unique four-digit integers with length equal to the number of choices in the whole quiz. For example, if there are 10 questions in the quiz and each question has 4 answer choices, \code{respids} will be a vector of length 10*4 = 40.}
}
\value{
Character string of the XML code for a single multiple-choice quiz question
}
\description{
Create the XML text for a single multiple-choice question. Takes as an argument a data frame with all the multiple choice questions of the quiz, all the response IDs for every response on the quiz, and an integer identifying which row in the data frame that the question appears.
It is necessary to include information for the whole quiz, because each question needs response IDs that are unique to the entire quiz.
}
