% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateQTI.R
\name{create_questionxml_essay}
\alias{create_questionxml_essay}
\title{Create the XML text for a single essage question}
\usage{
create_questionxml_essay(df, qn)
}
\arguments{
\item{df}{data.frame where each row is a question. The column names should include the following:
\itemize{
\item G: Question group
\item Question: Text of the question
\item Feedback: General feedback given to students after they complete the quiz and answers are shown
}}

\item{qn}{integer identifying the row in \code{df} that the individual question of interest appears}
}
\value{
Character string of the XML code for a single multiple-choice quiz question
}
\description{
Create the XML text for a single essay question. Takes as an argument a data frame with all the questions of the quiz and an integer identifying which row in the data frame that the question appears.
It is necessary to include information for the whole quiz, because some questions may need response IDs that are unique to the entire quiz.
}
