% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateQTI.R
\name{encode_equations}
\alias{encode_equations}
\title{Encode equations for use in URLs}
\usage{
encode_equations(latex_str, query_params = "scale=1")
}
\arguments{
\item{latex_str}{Character string of the LaTeX equation to be encoded}

\item{query_params}{Named list of query parameters to append to the encoded string. For example, \code{list(scale = 1)} will append \code{?scale=1} to the encoded string. Default is \code{list(scale = 1)}.}
}
\value{
Character string of the double URL-encoded LaTeX equation with appended query parameters if provided
}
\description{
Encode equations for use in URLs. This function performs two rounds of URL encoding to ensure that all special characters are properly encoded for use in URLs. The function also appends query parameters to the encoded string if provided.
}
