% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateQTI.R
\name{extract_equations}
\alias{extract_equations}
\title{Extract equations from a question string}
\usage{
extract_equations(question_str)
}
\arguments{
\item{question_str}{Character string of the question text, which may contain equations delimited with $[ ]$

[ ]: R:\%20}
}
\value{
Vector of character strings, where each string is a LaTeX equation extracted from the question string, without the $[ ]$ delimiters

[ ]: R:\%20
}
\description{
Extract equations from a question string. Equations are delimited with $[ ]$ in the question string. This function returns a vector of the equations without the delimiters.

[ ]: R:\%20
}
