% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateQTI.R
\name{extract_qtizip}
\alias{extract_qtizip}
\title{Unzip the QTI file and return the filepath for the XML file containing the assessment}
\usage{
extract_qtizip(filename, folder)
}
\arguments{
\item{filename}{Character string of the filename for the .zip file. Do not include the folder/path}

\item{folder}{Character string of the folder for the .zip file. The contents of the zip file will be extracted to the same folder.}
}
\value{
Character string of the filepath to the XML file containing the assessment tag (i.e. the XML file with the quiz questions)
}
\description{
This function unzips the QTI .zip file that is in \code{folder}, extracting the contents to the same folder, and returns the filepath for the XML file that contains the quiz questions
}
