% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateQTI.R
\name{extract_question}
\alias{extract_question}
\title{Extract information for a single quiz question from XML}
\usage{
extract_question(question.xml)
}
\arguments{
\item{question.xml}{The XMLNode object associated with a quiz question}
}
\value{
Returns a single-row tibble with the following columns:
\itemize{
\item G: Question group
\item 'Question Type': Character string identifying the type of question, should equal either 'MC' for a multiple choice question or 'Essay' for an essay/short-answer question
\item 'Text Type': Character string identifying whether the question text is plain text or html, should equal either 'html' or 'plain'
\item Points: Number of points for the question
\item Question: Text of the question
\item A: Correct answer, a number 1 thru 4
\item 'Choice 1': Choice 1
\item 'Choice 2': Choice 2
\item 'Choice 3': Choice 3
\item 'Choice 4': Choice 4
\item Feedback: General feedback given to students after they complete the quiz and answers are shown
}
}
\description{
Extract information for a single quiz question from an XML node and return in a single-row tibble
}
