% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateQTI.R
\name{extract_quiz_xml}
\alias{extract_quiz_xml}
\title{Extract all the questions from a quiz given in an XML file}
\usage{
extract_quiz_xml(xml_filepath)
}
\arguments{
\item{xml_filepath}{Character string equal to the file path for the XML file containing the assessment}
}
\value{
Returns a tibble with all the questions in the group, where each row is a question, and the tibble has the following columns:
\itemize{
\item G: Question group
\item 'Question Type': Character string identifying the type of question, should equal either 'MC' for a multiple choice question or 'Essay' for an essay/short-answer question
\item 'Text Type': Character string identifying whether the question text is plain text or html, should equal either 'html' or 'plain'
\item Points: Number of points for the question
\item Question: Text of the question
\item A: Correct answer, a number 1 thru 4
\item 'Choice 1': Choice 1
\item 'Choice 2': Choice 2
\item 'Choice 3': Choice 3
\item 'Choice 4': Choice 4
\item Feedback: General feedback given to students after they complete the quiz and answers are shown
}
}
\description{
Extract all the questions from a quiz given in an XML file and return in a tibble
}
