% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateQTI.R
\name{generateQTI}
\alias{generateQTI}
\title{Generate QTI quiz file from data frame}
\usage{
generateQTI(df, outfolder, quiztitle, quizfilename)
}
\arguments{
\item{df}{This is a tibble (i.e. data.frame) where each row is a question. The column names should be the following:
\itemize{
\item G: Question group
\item 'Question Type': Character string identifying the type of question, should equal either 'MC' for a multiple choice question or 'Essay' for an essay/short-answer question
\item 'Text Type': Character string identifying whether the question text is plain text or html, should equal either 'html' or 'plain'
\item Points: Number of points for the question
\item Question: Text of the question
\item A: Correct answer, a number 1 thru 4
\item 'Choice 1': Choice 1
\item 'Choice 2': Choice 2
\item 'Choice 3': Choice 3
\item 'Choice 4': Choice 4
\item Feedback: General feedback given to students after they complete the quiz and answers are shown
}}

\item{outfolder}{Character string with path to the output folder for the QTI quiz file}

\item{quiztitle}{Character string with the title for the quiz}

\item{quizfilename}{Character string with the filename for the quiz}
}
\description{
This function generates a zipped QTI file for a multiple-choice test given by the data frame, df, and saves it in the folder \code{outfolder}
}
