% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateQTI.R
\name{imgtag_equations}
\alias{imgtag_equations}
\title{Create the image tag for a LaTeX equation string}
\usage{
imgtag_equations(latex_eq_str)
}
\arguments{
\item{latex_eq_str}{Character string of the LaTeX equation to be converted into an image tag}
}
\value{
Character string of the image tag for the LaTeX equation, formatted for use in Canvas quizzes
}
\description{
Create the image tag for a LaTeX equation string. This function takes a LaTeX equation string, encodes it for use in a URL, and then creates the appropriate image tag for the equation using the codecogs equation image service. The resulting image tag is formatted for use in Canvas quizzes, with the appropriate attributes for accessibility and lazy loading.
}
