% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateQTI.R
\name{quiz_tibble}
\alias{quiz_tibble}
\title{Return a quiz tibble}
\usage{
quiz_tibble(nrows = 0)
}
\arguments{
\item{nrows}{Number of rows for the tibble. Default is 0.}
}
\value{
Returns a tibble (i.e. data.frame) where each row can hold a quiz question. The column names are those that are used in this package for quizzes
\itemize{
\item G: Question group
\item 'Question Type': Character string identifying the type of question, should equal either 'MC' for a multiple choice question or 'Essay' for an essay/short-answer question
\item 'Text Type': Character string identifying whether the question text is plain text or html, should equal either 'html' or 'plain'
\item Points: Number of points for the question
\item Question: Text of the question
\item A: Correct answer, a number 1 thru 4
\item 'Choice 1': Choice 1
\item 'Choice 2': Choice 2
\item 'Choice 3': Choice 3
\item 'Choice 4': Choice 4
\item Feedback: General feedback given to students after they complete the quiz and answers are shown
}
}
\description{
Return an empty tibble with the columns associated with quiz data frames used in this package
}
