% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_quiz.R
\name{read_quiz_googledoc}
\alias{read_quiz_googledoc}
\title{Read a quiz in from a Google Docs file and return a quiz data frame.}
\usage{
read_quiz_googledoc(doc_url, noauth = FALSE, tbl_number = 1, overwrite = TRUE)
}
\arguments{
\item{doc_url}{Path to the Google Docs file, with a table hopefully mostly conforming to the above.}

\item{noauth}{Logical, default is FALSE. Set equal to TRUE if the spreadsheet is accessible without Google login authorization and you would like to skip the login. Turns off authorization using \code{googledrive::deauth()}.}

\item{tbl_number}{If there is more than one table in the file, specify the table number you would like to import. Default is 1.}

\item{overwrite}{The function has the side effect of saving the Google Doc as a Word .docx file. If the .docx file already exists, set to TRUE or FALSE to overwrite the .docx file. The default is equal to TRUE.}
}
\value{
Quiz data frame with the columns above
}
\description{
This function reads a quiz in from a Google Docs file. The quiz must be in a table in the document. The function has the side effect of saving the Google Doc as a Word .docx file. The functions returns a quiz data frame that conforms to the following:
\itemize{
\item G: Question group
\item Question: Text of the question
\item Question type: Either MC or Essay, for multiple-choice or short-answer. The default is based on whether there is an answer given in column \code{A}, 'MC' if there is, 'Essay' otherwise.
\item Text Type: The type of text in all text columns, 'html' or 'plain', for html-formatted text or plain text. The default is 'html'.
\item Points: Numeric, number of points for the problem. The default is 1.
\item A: Correct answer for a multiple choice question, a number 1 thru 4. Default is NA.
\item 'Choice 1': Choice 1, default is empty character string ""
\item 'Choice 2': Choice 2, default is empty character string ""
\item 'Choice 3': Choice 3, default is empty character string ""
\item 'Choice 4': Choice 4, default is empty character string ""
\item Feedback: General feedback given to students after they complete the quiz and answers are shown. Default is empty character string "".
}

The only required column of the input data frame is \code{Question}. Any other missing columns will be created with default values

The function will also save a .docx version of the Google Doc in the default file path.
}
