% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_quiz.R
\name{read_quiz_json}
\alias{read_quiz_json}
\title{Read a quiz in from a JSON file and return a quiz data frame.}
\usage{
read_quiz_json(filepath)
}
\arguments{
\item{filepath}{Path to the JSON file}
}
\value{
Quiz data frame
}
\description{
This function reads a quiz in from a JSON file.
}
