% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_quiz.R
\name{render_pdf_quiz}
\alias{render_pdf_quiz}
\title{Render a TeX and PDF file of a quiz}
\usage{
render_pdf_quiz(
  quiz.df,
  outfile,
  outfolder,
  quiz_title,
  quiz_subtitle = "",
  instructor = "",
  includeanswers = TRUE,
  version = "a",
  seed = 1
)
}
\arguments{
\item{quiz.df}{data.frame where each row is a question. The column names should be the following:
\itemize{
\item G: Question group
\item Question: Text of the question
\item A: Correct answer, a number 1 thru 4
\item 'Choice 1': Choice 1
\item 'Choice 2': Choice 2
\item 'Choice 3': Choice 3
\item 'Choice 4': Choice 4
\item Feedback: General feedback given to students after they complete the quiz and answers are shown
}}

\item{outfile}{String with the name of the output file, but do not include the extension. File names with extensions .md, .tex, and .pdf will be created}

\item{outfolder}{String with the name of the folder to save the output}

\item{quiz_title}{String with the title of the quiz}

\item{quiz_subtitle}{String with a subtitle for the quiz, default is ""}

\item{instructor}{String with the instructor name, default is ""}

\item{includeanswers}{Boolean, set equal to TRUE to include answers in the output, FALSE to not include answers in the output. Default value is TRUE.}

\item{version}{Character, set to "a" or "b" for a version number. If there are more than two questions in any group of questions, for the name random number seed, the two versions will have two different questions. Default value is "a".}
}
\description{
This function takes a dataframe with a quiz and creates Markdown file for the quiz, and renders to TeX and PDF files
}
