% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateQTI.R
\name{replace_equations}
\alias{replace_equations}
\title{Create the image tags for all equations in a question string and replace the equations with the corresponding image tags}
\usage{
replace_equations(question_str)
}
\arguments{
\item{question_str}{Character string of the question text, which may contain equations delimited with $[ ]$

[ ]: R:\%20}
}
\value{
Character string of the question text with all equations replaced by their corresponding image tags
}
\description{
Create the image tags for all equations in a question string and replace the equations with the corresponding image tags. This function replaces all the equations from the question with an image tag for each equation
}
